WIP: bug fixes in fast posit<8,0> now it is using the C api

+++ c_api/pure_c/posit/posit8.c
@@ -14,7 +14,7 @@
 posit8_t posit8_sqrt(posit8_t a) {
 	posit8_t p;
 	float f = posit8_tof(a);
-	float root = sqrt(f);
+	float root = sqrtf(f);
 	p = posit8_fromf(root);
 	return p;
 }
